Fix a bug with the disk statistics in the Linux PMDA.  The sizes and signedness of the sprintf format and several of the fields of the per-device structure did not match the kernel, and did not cater for differences between 32 and 64 bit longs.
Initially observed as disk read/write_bytes and blkread/write on
one of our product NAS servers not being updated - this turned out
to be due to using signed vs unsigned integers in a sprintf.  The
fix was verified on the same machine:

~$ pminfo -f disk.dev.blkread

disk.dev.blkread
    inst [3 or "sdc"] value 2147483647
~$ gdb -q
(gdb) p /x 2147483647
$1 = 0x7fffffff

~$ pminfo -f disk.dev.blkread

disk.dev.blkread
    inst [3 or "sdc"] value 3247668682
$ gdb -q
(gdb) p /x 3247668682
$1 = 0xc1937dca

+++ src/pmdas/linux/proc_partitions.h
@@ -30,13 +30,13 @@ typedef struct {
     unsigned long	nr_blocks;
     char		*namebuf;     /* from /proc/{partitions,diskstats} */
     char		*udevnamebuf; /* from udev if we have it, else NULL */
-    unsigned int	rd_ios;
-    unsigned int	rd_merges;
-    unsigned int	rd_sectors;
+    unsigned long	rd_ios;
+    unsigned long	rd_merges;
+    unsigned long long	rd_sectors;
     unsigned int	rd_ticks;
-    unsigned int	wr_ios;
-    unsigned int	wr_merges;
-    unsigned int	wr_sectors;
+    unsigned long	wr_ios;
+    unsigned long	wr_merges;
+    unsigned long long	wr_sectors;
     unsigned int	wr_ticks;
     unsigned int	ios_in_flight;
     unsigned int	io_ticks;
